// online-store/src/interfaces/filter.interface.ts
export interface Producer {
    id: string;
    img: string;
}

export interface Camera {
    id: string;
    numberOfCameras: string;
}

export interface Color {
    id: string;
    color: string;
}

export interface Popular {
    id: string;
    value: boolean;
    isPopular: string;
}